# [Opt](mow) Calc delete bitmaps between segments concurrently in commit phase