--- a/be/src/common/config.h
+++ b/be/src/common/config.h
@@ -1566,6 +1566,8 @@ DECLARE_mInt32(load_trigger_compaction_version_percent);
 DECLARE_mInt64(base_compaction_interval_seconds_since_last_operation);
 DECLARE_mBool(enable_compaction_pause_on_high_memory);
 
+DECLARE_mBool(enable_calc_delete_bitmap_between_segments_concurrently);
+
 #ifdef BE_TEST
 // test s3
 DECLARE_String(test_s3_resource);