### What problem does this PR solve?

This PR make the calculation of delete bitmaps between segments in commit phase and the calculation of delete bitmaps of current load's rowsets with incremental historical data in commit phase concurrently for all tablets under one load channel. We use the initial load time(no historical data) of large dataset which can produce multi segments to evaluate the effectiveness of this PR.

test:

||210GB origin data files, 2 varchar(2000) as keys, cloud mode, 1 * FE 4C 16GB， 2 * BE 16C 128GB，PL2 ESSD*2, before|210GB origin data files, 2 varchar(2000) as keys, cloud mode, 1 * FE 4C 16GB， 2 * BE 16C 128GB，PL2 ESSD*2, with this PR|
|---|---|---|
|total time(s)|9 min 40 s|6 min 2 s|
|one load channel from beginning to last memtable flush|3 min 58 s|4 min 29 s|
|one load channel from last memtable flush to load channel close|5 min 40 s|1 min 25 s|

note that when calculating delete bitmaps between segments concurrently in commit phase, it requires high read iops and throughput

with this pr:
![image](https://github.com/user-attachments/assets/785355bc-dbf5-4ca2-b90c-270817dc4314)

without this pr:
![image](https://github.com/user-attachments/assets/676314ce-85cc-49a4-a393-b5de2781a838)

### Release note

None

### Check List (For Author)

- Test <!-- At least one of them must be included. -->
    - [ ] Regression test
    - [ ] Unit Test
    - [x] Manual test (add detailed scripts or steps below)
    - [ ] No need to test or manual test. Explain why:
        - [ ] This is a refactor/code format and no logic has been changed.
        - [ ] Previous test can cover this change.
        - [ ] No code files have been changed.
        - [ ] Other reason <!-- Add your reason?  -->

- Behavior changed:
    - [x] No.
    - [ ] Yes. <!-- Explain the behavior change -->

- Does this need documentation?
    - [x] No.
    - [ ] Yes. <!-- Add document PR link here. eg: https://github.com/apache/doris-website/pull/1214 -->

### Check List (For Reviewer who merge this PR)

- [ ] Confirm the release note
- [ ] Confirm test cases
- [ ] Confirm document
- [ ] Add branch pick label <!-- Add branch pick label that this PR should merge into -->