@hello-stephen 2025-04-16
# BE UT Coverage Report
Increment line coverage `0.00% (0/22)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/998bbf9fe84eccdae2cc7a425628f4196315b355_998bbf9fe84eccdae2cc7a425628f4196315b355/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/998bbf9fe84eccdae2cc7a425628f4196315b355_998bbf9fe84eccdae2cc7a425628f4196315b355/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 53.14% (14421/27136) |
| Line Coverage     | 42.01% (124995/297531) |
| Region Coverage   | 40.81% (63845/156454) |
| Branch Coverage   | 35.44% (32108/90594) |