@bobhan1 2025-04-21
run buildall