@bobhan1 2025-04-22
run buildall