@hello-stephen 2025-04-23
# BE Regression P0 && UT Coverage Report
Increment line coverage `0.00% (0/27)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/50106_bf56d6e187670765b2abe6ebf6f6eda3998a8efd_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/50106_bf56d6e187670765b2abe6ebf6f6eda3998a8efd_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage |                    |
| Line Coverage     |                    |
| Region Coverage   |                    |
| Branch Coverage   |                    |