@hello-stephen 2025-04-23
# BE UT Coverage Report
Increment line coverage `0.00% (0/27)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/613bd71ea4d16a11144ece7405db425a644ef9ab_613bd71ea4d16a11144ece7405db425a644ef9ab/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/613bd71ea4d16a11144ece7405db425a644ef9ab_613bd71ea4d16a11144ece7405db425a644ef9ab/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 53.54% (14540/27159) |
| Line Coverage     | 42.35% (126048/297614) |
| Region Coverage   | 41.18% (64427/156457) |
| Branch Coverage   | 35.76% (32386/90568) |