@hello-stephen 2025-04-24
# BE UT Coverage Report
Increment line coverage `0.00% (0/27)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/2ca33c5586d1aa955343c6495d7f6985adf4cea9_2ca33c5586d1aa955343c6495d7f6985adf4cea9/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/2ca33c5586d1aa955343c6495d7f6985adf4cea9_2ca33c5586d1aa955343c6495d7f6985adf4cea9/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 53.58% (14547/27148) |
| Line Coverage     | 42.42% (126263/297621) |
| Region Coverage   | 41.23% (64510/156482) |
| Branch Coverage   | 35.82% (32453/90588) |