@hello-stephen 2025-04-25
# BE Regression P0 && UT Coverage Report
Increment line coverage `0.00% (0/27)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/50106_2ca33c5586d1aa955343c6495d7f6985adf4cea9_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/50106_2ca33c5586d1aa955343c6495d7f6985adf4cea9_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 55.07% (14681/26658) |
| Line Coverage     | 44.53% (132301/297134) |
| Region Coverage   | 41.67% (76218/182903) |
| Branch Coverage   | 35.76% (36878/103140) |