@hello-stephen 2025-04-25
# BE Regression P0 && UT Coverage Report
Increment line coverage `0.00% (0/27)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/50106_2ca33c5586d1aa955343c6495d7f6985adf4cea9_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/50106_2ca33c5586d1aa955343c6495d7f6985adf4cea9_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 54.97% (14655/26658) |
| Line Coverage     | 44.41% (131957/297134) |
| Region Coverage   | 41.52% (75935/182903) |
| Branch Coverage   | 35.62% (36734/103140) |